refactor(R1.1): split repositories.py into db/repositories/ by entity
