+++ services/ingestion/db/repositories/analytics/__init__.py
@@ -1,0 +1,27 @@
+# services/ingestion/db/repositories/analytics/__init__.py
+"""Analytics repositories — re-export публічного API."""
+from services.ingestion.db.repositories.analytics.hero_stats import (
+    HeroRoleStatsRow,
+    HeroStatsRepository,
+    HeroStatsRow,
+)
+from services.ingestion.db.repositories.analytics.item_build import (
+    ItemBuildEntry,
+    ItemBuildRepository,
+)
+from services.ingestion.db.repositories.analytics.timeline import (
+    MatchPhaseStatsRow,
+    MatchTimelineRepository,
+    MetaHeroRow,
+)
+
+__all__ = [
+    "HeroStatsRepository",
+    "HeroStatsRow",
+    "HeroRoleStatsRow",
+    "ItemBuildRepository",
+    "ItemBuildEntry",
+    "MatchTimelineRepository",
+    "MatchPhaseStatsRow",
+    "MetaHeroRow",
+]
